+++ xhealth/log.go
@@ -4,102 +4,77 @@ import (
 	"fmt"
 
 	healthlog "github.com/InVisionApp/go-logger"
-	kitlog "github.com/go-kit/log"
-	"github.com/go-kit/log/level"
-	"github.com/xmidt-org/themis/xlog"
+	"go.uber.org/zap"
 )
 
 type loggerAdapter struct {
-	kitlog.Logger
+	Logger *zap.Logger
 }
 
 func (la loggerAdapter) Debug(msg ...interface{}) {
-	la.Logger.Log(
-		level.Key(), level.DebugValue(),
-		xlog.MessageKey(), fmt.Sprint(msg...),
-	)
+	la.Logger.Debug(fmt.Sprint(msg...))
 }
 
 func (la loggerAdapter) Info(msg ...interface{}) {
-	la.Logger.Log(
-		level.Key(), level.InfoValue(),
-		xlog.MessageKey(), fmt.Sprint(msg...),
-	)
+	la.Logger.Info(fmt.Sprint(msg...))
 }
 
 func (la loggerAdapter) Warn(msg ...interface{}) {
-	la.Logger.Log(
-		level.Key(), level.WarnValue(),
-		xlog.MessageKey(), fmt.Sprint(msg...),
-	)
+	la.Logger.Warn(fmt.Sprint(msg...))
 }
 
 func (la loggerAdapter) Error(msg ...interface{}) {
-	la.Logger.Log(
-		level.Key(), level.ErrorValue(),
-		xlog.MessageKey(), fmt.Sprint(msg...),
-	)
+	la.Logger.Error(fmt.Sprint(msg...))
 }
 
 func (la loggerAdapter) Debugln(msg ...interface{}) {
-	la.Debug(msg...)
+	la.Logger.Sugar().Debugln(msg...)
 }
 
 func (la loggerAdapter) Infoln(msg ...interface{}) {
-	la.Info(msg...)
+	la.Logger.Sugar().Infoln(msg...)
 }
 
 func (la loggerAdapter) Warnln(msg ...interface{}) {
-	la.Warn(msg...)
+	la.Logger.Sugar().Warnln(msg...)
 }
 
 func (la loggerAdapter) Errorln(msg ...interface{}) {
-	la.Error(msg...)
+	la.Logger.Sugar().Errorln(msg...)
 }
 
 func (la loggerAdapter) Debugf(format string, args ...interface{}) {
-	la.Logger.Log(
-		level.Key(), level.DebugValue(),
-		xlog.MessageKey(), fmt.Sprintf(format, args...),
-	)
+	la.Logger.Sugar().Debugf(format, args...)
+
 }
 
 func (la loggerAdapter) Infof(format string, args ...interface{}) {
-	la.Logger.Log(
-		level.Key(), level.InfoValue(),
-		xlog.MessageKey(), fmt.Sprintf(format, args...),
-	)
+	la.Logger.Sugar().Infof(format, args...)
+
 }
 
 func (la loggerAdapter) Warnf(format string, args ...interface{}) {
-	la.Logger.Log(
-		level.Key(), level.WarnValue(),
-		xlog.MessageKey(), fmt.Sprintf(format, args...),
-	)
+	la.Logger.Sugar().Warnf(format, args...)
 }
 
 func (la loggerAdapter) Errorf(format string, args ...interface{}) {
-	la.Logger.Log(
-		level.Key(), level.ErrorValue(),
-		xlog.MessageKey(), fmt.Sprintf(format, args...),
-	)
+	la.Logger.Sugar().Errorf(format, args...)
 }
 
 func (la loggerAdapter) WithFields(f healthlog.Fields) healthlog.Logger {
-	fields := make([]interface{}, 0, 2*len(f))
+	fields := make([]zap.Field, 0, 2*len(f))
 	for name, value := range f {
-		fields = append(fields, name, value)
+		fields = append(fields, zap.Any(name, value))
 	}
 
 	return loggerAdapter{
-		Logger: kitlog.With(
-			la.Logger,
+		Logger: la.Logger.With(
 			fields...,
 		),
 	}
 }
 
 // NewHealthLoggerAdapter adapts a go-kit logger onto the go-logger package's logger interface
-func NewHealthLoggerAdapter(logger kitlog.Logger) healthlog.Logger {
+func NewHealthLoggerAdapter(logger *zap.Logger) healthlog.Logger {
 	return loggerAdapter{Logger: logger}
 }